+++ micro_manager/micro_manager.py
@@ -17,20 +17,27 @@ import os
 import sys
 import time
 import inspect
-from copy import deepcopy
 from typing import Dict
 from warnings import warn
+from typing import Callable
 
 import numpy as np
+import time
+
 import precice
 
 import csv
 
 from .micro_manager_base import MicroManager
+
 from .adaptivity.global_adaptivity import GlobalAdaptivityCalculator
 from .adaptivity.local_adaptivity import LocalAdaptivityCalculator
+
 from .domain_decomposition import DomainDecomposer
+
 from .micro_simulation import create_simulation_class
+from .tools.logging_wrapper import Logger
+
 
 try:
     from .interpolation import Interpolation
@@ -51,11 +58,21 @@ class MicroManagerCoupling(MicroManager):
             Name of the JSON configuration file (provided by the user).
         """
         super().__init__(config_file)
+
+        self._logger = Logger(__name__, None, self._rank)
+
+        self._config.set_logger(self._logger)
         self._config.read_json_micro_manager()
-        # Define the preCICE Participant
-        self._participant = precice.Participant(
-            "Micro-Manager", self._config.get_config_file_name(), self._rank, self._size
-        )
+
+        # Data names of data to output to the snapshot database
+        self._write_data_names = self._config.get_write_data_names()
+
+        # Data names of data to read as input parameter to the simulations
+        self._read_data_names = self._config.get_read_data_names()
+
+        self._micro_dt = self._config.get_micro_dt()
+
+        self._is_micro_solve_time_required = self._config.write_micro_solve_time()
 
         self._macro_mesh_name = self._config.get_macro_mesh_name()
 
@@ -68,7 +85,7 @@ class MicroManagerCoupling(MicroManager):
         self._interpolate_crashed_sims = self._config.interpolate_crashed_micro_sim()
         if self._interpolate_crashed_sims:
             if Interpolation is None:
-                self._logger.info(
+                self._logger.log_info_one_rank(
                     "Interpolation is turned off as the required package is not installed."
                 )
                 self._interpolate_crashed_sims = False
@@ -83,15 +100,13 @@ class MicroManagerCoupling(MicroManager):
         self._is_adaptivity_on = self._config.turn_on_adaptivity()
 
         if self._is_adaptivity_on:
-            self._number_of_sims_for_adaptivity = 0
-
             self._data_for_adaptivity: Dict[str, np.ndarray] = dict()
-            self._adaptivity_type = self._config.get_adaptivity_type()
 
             self._adaptivity_data_names = self._config.get_data_for_adaptivity()
 
             # Names of macro data to be used for adaptivity computation
             self._adaptivity_macro_data_names = dict()
+
             # Names of micro data to be used for adaptivity computation
             self._adaptivity_micro_data_names = dict()
             for name, is_data_vector in self._adaptivity_data_names.items():
@@ -104,6 +119,17 @@ class MicroManagerCoupling(MicroManager):
                 self._config.is_adaptivity_required_in_every_implicit_iteration()
             )
             self._micro_sims_active_steps = None
+
+        self._adaptivity_output_n = self._config.get_adaptivity_output_n()
+        self._output_adaptivity_cpu_time = self._config.output_adaptivity_cpu_time()
+
+        # Define the preCICE Participant
+        self._participant = precice.Participant(
+            "Micro-Manager",
+            self._config.get_precice_config_file_name(),
+            self._rank,
+            self._size,
+        )
 
     # **************
     # Public methods
@@ -118,44 +144,27 @@ class MicroManagerCoupling(MicroManager):
         """
         t, n = 0, 0
         t_checkpoint, n_checkpoint = 0, 0
-        similarity_dists_cp = None
-        is_sim_active_cp = None
-        sim_is_associated_to_cp = None
         sim_states_cp = [None] * self._local_number_of_sims
         number_of_active_sims = []
+
+        micro_sim_solve = self._get_solve_variant()
 
         dt = min(self._participant.get_max_time_step_size(), self._micro_dt)
 
         if self._is_adaptivity_on:
-            similarity_dists = np.zeros(
-                (
-                    self._number_of_sims_for_adaptivity,
-                    self._number_of_sims_for_adaptivity,
-                )
-            )
-
-            # Start adaptivity calculation with all sims active
-            is_sim_active = np.array([True] * self._number_of_sims_for_adaptivity)
-
-            # Active sims do not have an associated sim
-            sim_is_associated_to = np.full(
-                (self._number_of_sims_for_adaptivity), -2, dtype=np.intc
-            )
-
             # If micro simulations have been initialized, compute adaptivity before starting the coupling
             if self._micro_sims_init:
-                (
-                    similarity_dists,
-                    is_sim_active,
-                    sim_is_associated_to,
-                ) = self._adaptivity_controller.compute_adaptivity(
+                self._logger.log_info_one_rank(
+                    "Micro simulations have been initialized, so adaptivity will be computed before the coupling begins."
+                )
+
+                self._adaptivity_controller.compute_adaptivity(
                     dt,
                     self._micro_sims,
-                    similarity_dists,
-                    is_sim_active,
-                    sim_is_associated_to,
                     self._data_for_adaptivity,
                 )
+
+        adaptivity_cpu_time = 0.0
 
         while self._participant.is_coupling_ongoing():
 
@@ -170,77 +179,36 @@ class MicroManagerCoupling(MicroManager):
 
                 if self._is_adaptivity_on:
                     if not self._adaptivity_in_every_implicit_step:
-                        (
-                            similarity_dists,
-                            is_sim_active,
-                            sim_is_associated_to,
-                        ) = self._adaptivity_controller.compute_adaptivity(
+                        start_time = time.process_time()
+                        self._adaptivity_controller.compute_adaptivity(
                             dt,
                             self._micro_sims,
-                            similarity_dists,
-                            is_sim_active,
-                            sim_is_associated_to,
                             self._data_for_adaptivity,
                         )
 
+                        end_time = time.process_time()
+
+                        adaptivity_cpu_time = end_time - start_time
+
                         # Only checkpoint the adaptivity configuration if adaptivity is computed
                         # once in every time window
-                        similarity_dists_cp = np.copy(similarity_dists)
-                        is_sim_active_cp = np.copy(is_sim_active)
-                        sim_is_associated_to_cp = np.copy(sim_is_associated_to)
+                        self._adaptivity_controller.write_checkpoint()
 
-                    if self._adaptivity_type == "local":
-                        active_sim_ids = np.where(is_sim_active)[0]
-                    elif self._adaptivity_type == "global":
-                        active_sim_ids = np.where(
-                            is_sim_active[
-                                self._global_ids_of_local_sims[
-                                    0
-                                ] : self._global_ids_of_local_sims[-1]
-                                + 1
-                            ]
-                        )[0]
+                        active_sim_ids = (
+                            self._adaptivity_controller.get_active_sim_ids()
+                        )
 
-                    for active_id in active_sim_ids:
-                        self._micro_sims_active_steps[active_id] += 1
+                        for active_id in active_sim_ids:
+                            self._micro_sims_active_steps[active_id] += 1
 
             micro_sims_input = self._read_data_from_precice(dt)
 
-            if self._is_adaptivity_on:
-                if self._adaptivity_in_every_implicit_step:
-                    (
-                        similarity_dists,
-                        is_sim_active,
-                        sim_is_associated_to,
-                    ) = self._adaptivity_controller.compute_adaptivity(
-                        dt,
-                        self._micro_sims,
-                        similarity_dists,
-                        is_sim_active,
-                        sim_is_associated_to,
-                        self._data_for_adaptivity,
-                    )
+            micro_sims_output, adaptivity_time = micro_sim_solve(micro_sims_input, dt)
 
-                    if self._adaptivity_type == "local":
-                        active_sim_ids = np.where(is_sim_active)[0]
-                    elif self._adaptivity_type == "global":
-                        active_sim_ids = np.where(
-                            is_sim_active[
-                                self._global_ids_of_local_sims[
-                                    0
-                                ] : self._global_ids_of_local_sims[-1]
-                                + 1
-                            ]
-                        )[0]
-
-                    for active_id in active_sim_ids:
-                        self._micro_sims_active_steps[active_id] += 1
-
-                micro_sims_output = self._solve_micro_simulations_with_adaptivity(
-                    micro_sims_input, is_sim_active, sim_is_associated_to, dt
-                )
-            else:
-                micro_sims_output = self._solve_micro_simulations(micro_sims_input, dt)
+            if self._output_adaptivity_cpu_time:
+                adaptivity_cpu_time += adaptivity_time
+                for i in range(self._local_number_of_sims):
+                    micro_sims_output[i]["adaptivity_cpu_time"] = adaptivity_cpu_time
 
             # Check if more than a certain percentage of the micro simulations have crashed and terminate if threshold is exceeded
             if self._interpolate_crashed_sims:
@@ -259,7 +227,7 @@ class MicroManagerCoupling(MicroManager):
                     )
 
                 if crash_ratio > self._crash_threshold:
-                    self._logger.info(
+                    self._logger.log_info_any_rank(
                         "{:.1%} of the micro simulations have crashed exceeding the threshold of {:.1%}. "
                         "Exiting simulation.".format(crash_ratio, self._crash_threshold)
                     )
@@ -267,11 +235,10 @@ class MicroManagerCoupling(MicroManager):
 
             self._write_data_to_precice(micro_sims_output)
 
-            t += dt  # increase internal time when time step is done.
-            n += 1  # increase counter
-            self._participant.advance(
-                dt
-            )  # notify preCICE that time step of size dt is complete
+            t += dt
+            n += 1
+
+            self._participant.advance(dt)
 
             # Revert micro simulations to their last checkpoints if required
             if self._participant.requires_reading_checkpoint():
@@ -283,26 +250,24 @@ class MicroManagerCoupling(MicroManager):
                 # If adaptivity is computed only once per time window, the states of sims need to be reset too
                 if self._is_adaptivity_on:
                     if not self._adaptivity_in_every_implicit_step:
-                        similarity_dists = np.copy(similarity_dists_cp)
-                        is_sim_active = np.copy(is_sim_active_cp)
-                        sim_is_associated_to = np.copy(sim_is_associated_to_cp)
+                        self._adaptivity_controller.read_checkpoint()
 
-            else:  # Time window has converged, now micro output can be generated
-                self._logger.info(
-                    "Micro simulations {} - {} have converged at t = {}".format(
-                        self._micro_sims[0].get_global_id(),
-                        self._micro_sims[-1].get_global_id(),
-                        t,
-                    )
-                )
-
+            if (
+                self._participant.is_time_window_complete()
+            ):  # Time window has converged, now micro output can be generated
                 if self._micro_sims_have_output:
                     if n % self._micro_n_out == 0:
                         for sim in self._micro_sims:
                             sim.output()
 
-                if self._is_adaptivity_on:
-                    number_of_active_sims.append(np.sum(is_sim_active))
+                if (
+                    self._is_adaptivity_on
+                    and n % self._adaptivity_output_n == 0
+                    and self._rank == 0
+                ):
+                    self._adaptivity_controller.log_metrics(n)
+
+                self._logger.log_info_one_rank("Time window {} converged.".format(n))
 
         self._participant.finalize()
 
@@ -336,7 +301,6 @@ class MicroManagerCoupling(MicroManager):
         ), "Provided macro mesh bounds are of incorrect dimension"
         if self._is_parallel:
             domain_decomposer = DomainDecomposer(
-                self._logger,
                 self._participant.get_mesh_dimensions(self._macro_mesh_name),
                 self._rank,
                 self._size,
@@ -361,13 +325,10 @@ class MicroManagerCoupling(MicroManager):
         assert self._mesh_vertex_coords.size != 0, "Macro mesh has no vertices."
 
         self._local_number_of_sims, _ = self._mesh_vertex_coords.shape
-        self._logger.info(
-            "Number of local micro simulations = {}".format(self._local_number_of_sims)
-        )
 
         if self._local_number_of_sims == 0:
             if self._is_parallel:
-                self._logger.info(
+                self._logger.log_info_any_rank(
                     "Rank {} has no micro simulations and hence will not do any computation.".format(
                         self._rank
                     )
@@ -382,8 +343,35 @@ class MicroManagerCoupling(MicroManager):
         # the correct global IDs
         self._comm.Allgatherv(np.array(self._local_number_of_sims), nms_all_ranks)
 
+        max_nms = np.max(nms_all_ranks)
+        min_nms = np.min(nms_all_ranks)
+
+        if (
+            max_nms != min_nms
+        ):  # if the number of maximum and minimum micro simulations per rank are different
+            self._logger.log_info_one_rank(
+                "The following ranks have the maximum number of micro simulations ({}): {}".format(
+                    max_nms, np.where(nms_all_ranks == max_nms)[0]
+                )
+            )
+            self._logger.log_info_one_rank(
+                "The following ranks have the minimum number of micro simulations ({}): {}".format(
+                    min_nms, np.where(nms_all_ranks == min_nms)[0]
+                )
+            )
+        else:  # if the number of maximum and minimum micro simulations per rank are the same
+            self._logger.log_info_one_rank(
+                "All ranks have the same number of micro simulations: {}".format(
+                    max_nms
+                )
+            )
+
         # Get global number of micro simulations
-        self._global_number_of_sims = np.sum(nms_all_ranks)
+        self._global_number_of_sims: int = np.sum(nms_all_ranks)
+
+        self._logger.log_info_one_rank(
+            "Total number of micro simulations: {}".format(self._global_number_of_sims)
+        )
 
         if self._is_adaptivity_on:
             for name, is_data_vector in self._adaptivity_data_names.items():
@@ -408,8 +396,6 @@ class MicroManagerCoupling(MicroManager):
             self._global_ids_of_local_sims.append(sim_id)
             sim_id += 1
 
-        self._micro_sims = [None] * self._local_number_of_sims  # DECLARATION
-
         # Setup for simulation crashes
         self._has_sim_crashed = [False] * self._local_number_of_sims
         if self._interpolate_crashed_sims:
@@ -423,35 +409,35 @@ class MicroManagerCoupling(MicroManager):
         )
 
         # Create micro simulation objects
+        self._micro_sims = []
         for i in range(self._local_number_of_sims):
-            self._micro_sims[i] = create_simulation_class(micro_problem)(
-                self._global_ids_of_local_sims[i]
+            self._micro_sims.append(
+                create_simulation_class(micro_problem)(
+                    self._global_ids_of_local_sims[i]
+                )
             )
-
-        self._logger.info(
-            "Micro simulations with global IDs {} - {} created.".format(
-                self._global_ids_of_local_sims[0], self._global_ids_of_local_sims[-1]
-            )
-        )
 
         if self._is_adaptivity_on:
-            if self._adaptivity_type == "local":
-                self._adaptivity_controller = LocalAdaptivityCalculator(
-                    self._config, self._logger
+            if self._config.get_adaptivity_type() == "local":
+                self._adaptivity_controller: LocalAdaptivityCalculator = (
+                    LocalAdaptivityCalculator(
+                        self._config, self._rank, self._comm, self._local_number_of_sims
+                    )
                 )
-                self._number_of_sims_for_adaptivity = self._local_number_of_sims
-            elif self._adaptivity_type == "global":
-                self._adaptivity_controller = GlobalAdaptivityCalculator(
-                    self._config,
-                    self._logger,
-                    self._global_number_of_sims,
-                    self._global_ids_of_local_sims,
-                    self._rank,
-                    self._comm,
+            elif self._config.get_adaptivity_type() == "global":
+                self._adaptivity_controller: GlobalAdaptivityCalculator = (
+                    GlobalAdaptivityCalculator(
+                        self._config,
+                        self._global_number_of_sims,
+                        self._global_ids_of_local_sims,
+                        self._rank,
+                        self._comm,
+                    )
                 )
-                self._number_of_sims_for_adaptivity = self._global_number_of_sims
 
-            self._micro_sims_active_steps = np.zeros(self._local_number_of_sims)
+            self._micro_sims_active_steps = np.zeros(
+                self._local_number_of_sims
+            )  # DECLARATION
 
         self._micro_sims_init = False  # DECLARATION
 
@@ -485,7 +471,7 @@ class MicroManagerCoupling(MicroManager):
                         "The initialize() method of the Micro simulation has an incorrect number of arguments."
                     )
             except TypeError:
-                self._logger.info(
+                self._logger.log_info_one_rank(
                     "The signature of initialize() method of the micro simulation cannot be determined. Trying to determine the signature by calling the method."
                 )
                 # Try to get the signature of the initialize() method, if it is not written in Python
@@ -493,7 +479,7 @@ class MicroManagerCoupling(MicroManager):
                     self._micro_sims[0].initialize()
                     is_initial_data_required = False
                 except TypeError:
-                    self._logger.info(
+                    self._logger.log_info_one_rank(
                         "The initialize() method of the micro simulation has arguments. Attempting to call it again with initial data."
                     )
                     try:  # Try to call the initialize() method with initial data
@@ -653,7 +639,7 @@ class MicroManagerCoupling(MicroManager):
                     self._macro_mesh_name, dname, [], np.array([])
                 )
 
-    def _solve_micro_simulations(self, micro_sims_input: list, dt: float) -> list:
+    def _solve_micro_simulations(self, micro_sims_input: list, dt: float) -> tuple:
         """
         Solve all micro simulations and assemble the micro simulations outputs in a list of dicts format.
 
@@ -671,31 +657,31 @@ class MicroManagerCoupling(MicroManager):
             List of dicts in which keys are names of data and the values are the data of the output of the micro
             simulations.
         """
-        micro_sims_output = [None] * self._local_number_of_sims
+        micro_sims_output: list[dict] = [None] * self._local_number_of_sims
 
         for count, sim in enumerate(self._micro_sims):
             # If micro simulation has not crashed in a previous iteration, attempt to solve it
             if not self._has_sim_crashed[count]:
                 # Attempt to solve the micro simulation
                 try:
-                    start_time = time.time()
+                    start_time = time.process_time()
                     micro_sims_output[count] = sim.solve(micro_sims_input[count], dt)
-                    end_time = time.time()
+                    end_time = time.process_time()
                     # Write solve time of the macro simulation if required and the simulation has not crashed
                     if self._is_micro_solve_time_required:
-                        micro_sims_output[count]["micro_sim_time"] = (
+                        micro_sims_output[count]["solve_cpu_time"] = (
                             end_time - start_time
                         )
 
                 # If simulation crashes, log the error and keep the output constant at the previous iteration's output
                 except Exception as error_message:
-                    self._logger.error(
-                        "Micro simulation at macro coordinates {} has experienced an error. "
+                    self._logger.log_error_any_rank(
+                        "Micro simulation at macro coordinates {} with input {} has experienced an error. "
                         "See next entry on this rank for error message.".format(
-                            self._mesh_vertex_coords[count]
+                            self._mesh_vertex_coords[count], micro_sims_input[count]
                         )
                     )
-                    self._logger.error(error_message)
+                    self._logger.log_error_any_rank(error_message)
                     self._has_sim_crashed[count] = True
 
         # If interpolate is off, terminate after crash
@@ -705,7 +691,9 @@ class MicroManagerCoupling(MicroManager):
                 np.sum(self._has_sim_crashed), crashed_sims_on_all_ranks
             )
             if sum(crashed_sims_on_all_ranks) > 0:
-                self._logger.info("Exiting simulation after micro simulation crash.")
+                self._logger.log_info_any_rank(
+                    "Exiting simulation after micro simulation crash."
+                )
                 sys.exit()
 
         # Interpolate result for crashed simulation
@@ -716,7 +704,7 @@ class MicroManagerCoupling(MicroManager):
         # Iterate over all crashed simulations to interpolate output
         if self._interpolate_crashed_sims:
             for unset_sim in unset_sims:
-                self._logger.info(
+                self._logger.log_info_any_rank(
                     "Interpolating output for crashed simulation at macro vertex {}.".format(
                         self._mesh_vertex_coords[unset_sim]
                     )
@@ -725,27 +713,19 @@ class MicroManagerCoupling(MicroManager):
                     micro_sims_input, micro_sims_output, unset_sim
                 )
 
-        return micro_sims_output
+        return micro_sims_output, 0.0
 
     def _solve_micro_simulations_with_adaptivity(
-        self,
-        micro_sims_input: list,
-        is_sim_active: np.ndarray,
-        sim_is_associated_to: np.ndarray,
-        dt: float,
-    ) -> list:
+        self, micro_sims_input: list, dt: float
+    ) -> tuple:
         """
-        Solve all micro simulations and assemble the micro simulations outputs in a list of dicts format.
+        Adaptively solve micro simulations and assemble the micro simulations outputs in a list of dicts format.
 
         Parameters
         ----------
         micro_sims_input : list
             List of dicts in which keys are names of data and the values are the data which are required inputs to
             solve a micro simulation.
-        is_sim_active : numpy array
-            1D array having state (active or inactive) of each micro simulation
-        sim_is_associated_to : numpy array
-            1D array with values of associated simulations of inactive simulations. Active simulations have None
         dt : float
             Time step size.
 
@@ -755,27 +735,26 @@ class MicroManagerCoupling(MicroManager):
             List of dicts in which keys are names of data and the values are the data of the output of the micro
             simulations.
         """
-        if self._adaptivity_type == "global":
-            active_sim_ids = np.where(
-                is_sim_active[
-                    self._global_ids_of_local_sims[0] : self._global_ids_of_local_sims[
-                        -1
-                    ]
-                    + 1
-                ]
-            )[0]
-            inactive_sim_ids = np.where(
-                is_sim_active[
-                    self._global_ids_of_local_sims[0] : self._global_ids_of_local_sims[
-                        -1
-                    ]
-                    + 1
-                ]
-                == False
-            )[0]
-        elif self._adaptivity_type == "local":
-            active_sim_ids = np.where(is_sim_active)[0]
-            inactive_sim_ids = np.where(is_sim_active == False)[0]
+        adaptivity_cpu_time = 0.0
+
+        if self._adaptivity_in_every_implicit_step:
+            start_time = time.process_time()
+            self._adaptivity_controller.compute_adaptivity(
+                dt,
+                self._micro_sims,
+                self._data_for_adaptivity,
+            )
+            end_time = time.process_time()
+
+            adaptivity_cpu_time = end_time - start_time
+
+            active_sim_ids = self._adaptivity_controller.get_active_sim_ids()
+
+            for active_id in active_sim_ids:
+                self._micro_sims_active_steps[active_id] += 1
+
+        active_sim_ids = self._adaptivity_controller.get_active_sim_ids()
+        inactive_sim_ids = self._adaptivity_controller.get_inactive_sim_ids()
 
         micro_sims_output = [None] * self._local_number_of_sims
 
@@ -783,16 +762,15 @@ class MicroManagerCoupling(MicroManager):
         for active_id in active_sim_ids:
             # If micro simulation has not crashed in a previous iteration, attempt to solve it
             if not self._has_sim_crashed[active_id]:
-                # Attempt to solve the micro simulation
                 try:
-                    start_time = time.time()
+                    start_time = time.process_time()
                     micro_sims_output[active_id] = self._micro_sims[active_id].solve(
                         micro_sims_input[active_id], dt
                     )
-                    end_time = time.time()
+                    end_time = time.process_time()
                     # Write solve time of the macro simulation if required and the simulation has not crashed
                     if self._is_micro_solve_time_required:
-                        micro_sims_output[active_id]["micro_sim_time"] = (
+                        micro_sims_output[active_id]["solve_cpu_time"] = (
                             end_time - start_time
                         )
 
@@ -804,13 +782,13 @@ class MicroManagerCoupling(MicroManager):
 
                 # If simulation crashes, log the error and keep the output constant at the previous iteration's output
                 except Exception as error_message:
-                    self._logger.error(
+                    self._logger.log_error_any_rank(
                         "Micro simulation at macro coordinates {} has experienced an error. "
                         "See next entry on this rank for error message.".format(
                             self._mesh_vertex_coords[active_id]
                         )
                     )
-                    self._logger.error(error_message)
+                    self._logger.log_error_any_rank(error_message)
                     self._has_sim_crashed[active_id] = True
 
         # If interpolate is off, terminate after crash
@@ -820,8 +798,11 @@ class MicroManagerCoupling(MicroManager):
                 np.sum(self._has_sim_crashed), crashed_sims_on_all_ranks
             )
             if sum(crashed_sims_on_all_ranks) > 0:
-                self._logger.info("Exiting simulation after micro simulation crash.")
+                self._logger.log_error_any_rank(
+                    "Exiting simulation after micro simulation crash."
+                )
                 sys.exit()
+
         # Interpolate result for crashed simulation
         unset_sims = []
         for active_id in active_sim_ids:
@@ -831,7 +812,7 @@ class MicroManagerCoupling(MicroManager):
         # Iterate over all crashed simulations to interpolate output
         if self._interpolate_crashed_sims:
             for unset_sim in unset_sims:
-                self._logger.info(
+                self._logger.log_info_any_rank(
                     "Interpolating output for crashed simulation at macro vertex {}.".format(
                         self._mesh_vertex_coords[unset_sim]
                     )
@@ -841,16 +822,13 @@ class MicroManagerCoupling(MicroManager):
                     micro_sims_input, micro_sims_output, unset_sim, active_sim_ids
                 )
 
-        # For each inactive simulation, copy data from most similar active simulation
-        if self._adaptivity_type == "global":
-            self._adaptivity_controller.communicate_micro_output(
-                is_sim_active, sim_is_associated_to, micro_sims_output
-            )
-        elif self._adaptivity_type == "local":
-            for inactive_id in inactive_sim_ids:
-                micro_sims_output[inactive_id] = deepcopy(
-                    micro_sims_output[sim_is_associated_to[inactive_id]]
-                )
+        start_time = time.process_time()
+        micro_sims_output = self._adaptivity_controller.get_full_field_micro_output(
+            micro_sims_output
+        )
+        end_time = time.process_time()
+
+        adaptivity_cpu_time += end_time - start_time
 
         # Resolve micro sim output data for inactive simulations
         for inactive_id in inactive_sim_ids:
@@ -860,14 +838,30 @@ class MicroManagerCoupling(MicroManager):
             ] = self._micro_sims_active_steps[inactive_id]
 
             if self._is_micro_solve_time_required:
-                micro_sims_output[inactive_id]["micro_sim_time"] = 0
+                micro_sims_output[inactive_id]["solve_cpu_time"] = 0
 
         # Collect micro sim output for adaptivity calculation
         for i in range(self._local_number_of_sims):
             for name in self._adaptivity_micro_data_names:
                 self._data_for_adaptivity[name][i] = micro_sims_output[i][name]
 
-        return micro_sims_output
+        return micro_sims_output, adaptivity_cpu_time
+
+    def _get_solve_variant(self) -> Callable[[list, float], list]:
+        """
+        Get the solve variant function based on the adaptivity type.
+
+        Returns
+        -------
+        solve_variant : Callable
+            Solve variant function based on the adaptivity type.
+        """
+        if self._is_adaptivity_on:
+            solve_variant = self._solve_micro_simulations_with_adaptivity
+        else:
+            solve_variant = self._solve_micro_simulations
+
+        return solve_variant
 
     def _interpolate_output_for_crashed_sim(
         self,
@@ -926,7 +920,7 @@ class MicroManagerCoupling(MicroManager):
                 micro_sims_active_values.append(micro_sims_output[i].copy())
         # Find nearest neighbors
         if len(micro_sims_active_input_lists) == 0:
-            self._logger.error(
+            self._logger.log_error_any_rank(
                 "No active neighbors available for interpolation at macro vertex {}. Value cannot be interpolated".format(
                     self._mesh_vertex_coords[unset_sim]
                 )
@@ -947,13 +941,13 @@ class MicroManagerCoupling(MicroManager):
             if self._is_adaptivity_on:
                 interpol_space.append(micro_sims_active_input_lists[neighbor].copy())
                 interpol_values.append(micro_sims_active_values[neighbor].copy())
-                interpol_values[-1].pop("micro_sim_time", None)
+                interpol_values[-1].pop("solve_cpu_time", None)
                 interpol_values[-1].pop("active_state", None)
                 interpol_values[-1].pop("active_steps", None)
             else:
                 interpol_space.append(micro_sims_active_input_lists[neighbor].copy())
                 interpol_values.append(micro_sims_active_values[neighbor].copy())
-                interpol_values[-1].pop("micro_sim_time", None)
+                interpol_values[-1].pop("solve_cpu_time", None)
 
         # Interpolate for each parameter
         output_interpol = dict()
@@ -967,7 +961,7 @@ class MicroManagerCoupling(MicroManager):
             )
         # Reintroduce removed information
         if self._is_micro_solve_time_required:
-            output_interpol["micro_sim_time"] = 0
+            output_interpol["solve_cpu_time"] = 0
         if self._is_adaptivity_on:
             output_interpol["active_state"] = 1
             output_interpol["active_steps"] = self._micro_sims_active_steps[unset_sim]