+++ tests/unit/test_adaptivity_parallel.py
@@ -30,8 +30,10 @@ class TestGlobalAdaptivity(TestCase):
 
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
         # Force the activation of sim #0 and #4
@@ -93,9 +95,13 @@ class TestGlobalAdaptivity(TestCase):
             global_ids = [3, 4]
             data_for_adaptivity = {"data1": [1.0, 1.0], "data2": [13.0, 13.0]}
 
-        similarity_dists = np.zeros((5, 5))
-        is_sim_active = np.array([True, True, True, True, True])
-        sim_is_associated_to = [-2, -2, -2, -2, -2]
+        adaptivity_data = []
+        adaptivity_data.append(np.zeros((5, 5)))  # similarity_dists
+        adaptivity_data.append(
+            np.array([True, True, True, True, True])
+        )  # is_sim_active
+        adaptivity_data.append([-2, -2, -2, -2, -2])  # sim_is_associated_to
+
         expected_is_sim_active = np.array([False, False, False, False, True])
         expected_sim_is_associated_to = [4, 4, 4, 4, -2]
 
@@ -104,8 +110,10 @@ class TestGlobalAdaptivity(TestCase):
         configurator.get_adaptivity_refining_const = MagicMock(return_value=0.05)
         configurator.get_adaptivity_coarsening_const = MagicMock(return_value=0.2)
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L2rel")
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
         adaptivity_controller._adaptivity_data_names = {
@@ -132,22 +140,20 @@ class TestGlobalAdaptivity(TestCase):
         for i in global_ids:
             dummy_micro_sims.append(MicroSimulation(i))
 
-        (
-            _,
-            is_sim_active,
-            sim_is_associated_to,
-        ) = adaptivity_controller.compute_adaptivity(
+        adaptivity_controller.compute_adaptivity(
             0.1,
             dummy_micro_sims,
-            similarity_dists,
-            is_sim_active,
-            sim_is_associated_to,
             data_for_adaptivity,
         )
 
-        self.assertTrue(np.array_equal(expected_is_sim_active, is_sim_active))
         self.assertTrue(
-            np.array_equal(expected_sim_is_associated_to, sim_is_associated_to)
+            np.array_equal(expected_is_sim_active, adaptivity_controller._is_sim_active)
+        )
+        self.assertTrue(
+            np.array_equal(
+                expected_sim_is_associated_to,
+                adaptivity_controller._sim_is_associated_to,
+            )
         )
 
     def test_communicate_micro_output(self):
@@ -167,16 +173,18 @@ class TestGlobalAdaptivity(TestCase):
             sim_output = [output_1, None]
             expected_sim_output = [output_1, output_0]
 
-        is_sim_active = np.array([False, False, True, True, False])
-        sim_is_associated_to = [3, 3, -2, -2, 2]
+        is_sim_active = np.array([False, False, True, True, False])  # is_sim_active
+        sim_is_associated_to = [3, 3, -2, -2, 2]  # sim_is_associated_to
 
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
-        adaptivity_controller.communicate_micro_output(
+        adaptivity_controller._communicate_micro_output(
             is_sim_active, sim_is_associated_to, sim_output
         )