+++ tests/unit/test_adaptivity_serial.py
@@ -68,7 +68,9 @@ class TestLocalAdaptivity(TestCase):
         """
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
-        adaptivity_controller = AdaptivityCalculator(configurator, logger=MagicMock())
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
+        adaptivity_controller = AdaptivityCalculator(configurator, 0)
         adaptivity_controller._hist_param = 0.5
         adaptivity_controller._adaptivity_data_names = [
             "micro-scalar-data",
@@ -102,7 +104,9 @@ class TestLocalAdaptivity(TestCase):
         """
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
-        adaptivity_controller = AdaptivityCalculator(configurator, logger=MagicMock())
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
+        adaptivity_controller = AdaptivityCalculator(configurator, 0)
         adaptivity_controller._refine_const = self._refine_const
         adaptivity_controller._coarse_const = self._coarse_const
         adaptivity_controller._adaptivity_data_names = [
@@ -127,8 +131,7 @@ class TestLocalAdaptivity(TestCase):
         """
         Test functionality for calculating similarity criteria between pairs of simulations using different norms in class AdaptivityCalculator.
         """
-        logger = MagicMock()
-        calc = AdaptivityCalculator(Config(logger, "micro-manager-config.json"), logger)
+        calc = AdaptivityCalculator(Config("micro-manager-config.json"), 0)
 
         fake_data = np.array([[1], [2], [3]])
         self.assertTrue(
@@ -209,7 +212,9 @@ class TestLocalAdaptivity(TestCase):
         """
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
-        adaptivity_controller = AdaptivityCalculator(configurator, logger=MagicMock())
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
+        adaptivity_controller = AdaptivityCalculator(configurator, 0)
         adaptivity_controller._refine_const = self._refine_const
         adaptivity_controller._coarse_const = self._coarse_const
         adaptivity_controller._adaptivity_data_names = [
@@ -236,8 +241,10 @@ class TestLocalAdaptivity(TestCase):
         """
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
+        configurator.get_output_dir = MagicMock(return_value="output_dir")
+
         adaptivity_controller = LocalAdaptivityCalculator(
-            configurator, logger=MagicMock()
+            configurator, 0, MagicMock(), 5
         )
         adaptivity_controller._refine_const = self._refine_const
         adaptivity_controller._coarse_const = self._coarse_const